Removing explicit cast to float in getBorderWidth().
git-svn-id: http://chianti.ucsd.edu/svn@3590 0ecc0d97-ab19-0410-9704-bfe1a75892f5

+++ csplugins/trunk/ucsd/nlandys/fung/src/cytoscape/fung/DefaultNodeDetails.java
@@ -31,7 +31,7 @@ class DefaultNodeDetails extends NodeDetails
 
   public float borderWidth(int node)
   {
-    return (float) m_fung.m_nodeDefaults.m_borderWidth;
+    return m_fung.m_nodeDefaults.m_borderWidth;
   }
 
   public Paint borderPaint(int node)
